refactor: move model-spec module into the compile package _shared directory
The `ModelSpec` type and the `normalizeModelSpec` helper are shared across the
model reading and code generation steps, so they belong alongside the other
shared modules rather than at the top level of `src`.
