+++ packages/compile/README.md
@@ -35,7 +35,7 @@ The `compile` package is currently treated as an implementation detail of the `c
 As such, there is no generated API documentation at this time, but we hope to expose a public API once the interfaces stabilize.
 
 In the meantime, the types for the public API are declared using JSDoc comments alongside the code that defines them, and the `build` script generates the corresponding `.d.ts` files (under `dist`) that are published with the package.
-Of particular note, the [`ModelSpec`](./src/model-spec.js) type is the authoritative definition of the JSON "spec file" format that is passed to the `sde generate` command using the `--spec` argument, and every supported property is documented there.
+Of particular note, the [`ModelSpec`](./src/_shared/model-spec.js) type is the authoritative definition of the JSON "spec file" format that is passed to the `sde generate` command using the `--spec` argument, and every supported property is documented there.
 For a summary of those properties in prose form, see the ["Configuration files"](../cli/README.md#configuration-files) section of the `cli` package README.
 
 Note that the `ModelSpec` type in the [`@sdeverywhere/build`](../build/README.md) package (which describes the `modelSpec` section of an `sde.config.js` file) is derived from the `ModelSpec` type in this package, so the two formats share a single definition for the properties they have in common.
